Refactor SSSOM Schema usage
This reduces other places where the instantiation of the schema happens
